requirements.txt
Flask==3.1
Flask-Cors==3.0.10
Werkzeug==3.1
tensorflow-cpu==2.18.1
matplotlib==3.7
numpy==2.0
gunicorn==20.1
